fix: #tno-2269 - fixed reporting service - code to include aggregations on section was missing - moved converter to shared library
